Add Kelvin wave test case
This commit adds the test case of a Kelvin wave in a 2D periodic
channel.  The example shows that the wave propagates with the correct
phase speed as predicted by linear wave theory.  There appear some small
irregularities, which we might come from non-perfect initial conditions
or might show the limits of linear wave theory.